2024 day 7
not only did i start late (~12:40) but my dumb ass decided that it would be a good idea to reinvent the binary counter, which then led to the trinary(??) counter in part 2. pain. horrible pain. i was literally over a hundred trillion off on my first attempt because my code didn't function. aoc weekends smh.

also my code for part 2 is SLOW. at least i can still measure the runtime in single digit seconds, but it's doing so many calculation omg. is there even a way to do this that _isn't_ bruteforce?

part 1 - rank 9851 (1:25:49 -> ~00:45)
part 2 - rank 10376 (2:06:27 -> ~1:26)

crazy how i took 45 minutes to do part 2 and still was that close in rank to part 1.