2024 day 11 part 2
would you look at that i did it. my dumbass self originally rewrote this with nested for loops instead of the `+ stones[odds[j]]` and `+ stones[evens[j]]`, so it was literally no improvement. once i figured it out it solved it in less than a second. fun. wow i feel so dumb.

part 2: rank 7885 (2:00:30)

i could've done this an hour ago. why didn't i do this an hour ago?????