2024 day 14
i had to find out what a tree looked like from reddit because the puzzle text is extremely vague. all it had to say was "the robots should arrange themselves into a *small* picture of a christmas tree in the *center of the lobby*" and it would've been enough (even without specifying it's small it would've been fine).

i ended up rewriting a lot of part 2 because for my solve i just bruteforced 10000 and returned the one with points the lowest average distance from the center.

i like how my code turned out, even if i did waste an hour trying to make it look pretty.

part 1: rank 4281 (37:09)
part 2: rank 3948 (1:21:00)