2024 day 15 part 2
SIKE IM NOT LETTING A STUPID CALENDAR GET THE BEST OF ME

part 2: rank 5180 (3:45:34)

i got the _example_ working at around 3:30, and it took me 15 minutes to finish the rest. the variable and function names are nonsensical because they're mixed and matched from ~3 different attempts. this last one worked. IT WORKED!

THE STREAK LIVES ON ANOTHER DAY AND MY SLEEP DEPRIVATION TOO HAHAHAHAHA THIS IS FUN