2025 day 9 part 1
part 1: 04:08
rank somewhere between 231-2288

finished part 1 in 4 minutes, and naturally here i am 2 and a half hours later aimlessly working on part 2. i can't even get it working for the tests, let alone on the actual input (i know i'd have to optimize it). so im done. im not coming back to this. geometry and pathfinding is always my downfall.