2025 day 10 part 1
part 1: 34:18
rank somewhere between 2000-3500

spent 3 hours on part 2. i have a final coming up. yay life choices.
i hate when my solution works but the input is just too big for it to be practical. it was supposed to like iterate through every combination of sums or something but that's too slow for bigger numbers like this.